+++ packages/api-auth-entity-storage-service/src/restEntryPoints.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IRestRouteEntryPoint } from "@gtsc/api-models";
+import type { IRestRouteEntryPoint } from "@twin.org/api-models";
 import {
 	generateRestRoutesAuthentication,
 	tagsAuthentication