+++ packages/api-models/src/models/protocol/IHttpResponse.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { HttpStatusCode, IHttpHeaders } from "@gtsc/web";
+import type { HttpStatusCode, IHttpHeaders } from "@twin.org/web";
 
 /**
  * Model for the standard parameters for an http response.