+++ packages/api-models/src/models/responses/errors/IForbiddenResponse.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IError } from "@gtsc/core";
-import type { HttpStatusCode } from "@gtsc/web";
+import type { IError } from "@twin.org/core";
+import type { HttpStatusCode } from "@twin.org/web";
 
 /**
  * The operation that you tried to perform is not possible, see the content for more details.