+++ packages/api-models/tests/models/routes/IRestRoute.spec.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { nameof } from "@gtsc/nameof";
+import { nameof } from "@twin.org/nameof";
 import type { IHttpRequestContext } from "../../../src/models/protocol/IHttpRequestContext";
 import type { INotFoundResponse } from "../../../src/models/responses/errors/INotFoundResponse";
 import type { IRestRoute } from "../../../src/models/routes/IRestRoute";