+++ packages/api-processors/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/api-processors - Changelog
+# @twin.org/api-processors - Changelog
 
-## v0.0.3
+## v0.0.1-next.2
 
 - Initial Release